softfork: new module currently to track softforks.
The structure is called "BIP" for brevity, even though it could be any
soft fork.  Currently it has simple stateless logic to support versionbits
BIPs (BIP9).

We keep pointers to the version BIPs (globals) in a table: once there
are many this will make it clear what order they resolve in.

The BIP class has only "IsActive()" which can be trivially made to
work for IsSuperMajority()-style activation as well.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>