UpdateTip: hook in softfork check.
While we're not currently monitoring any BIPs, we need this so we
are warned if an unknown softfork appears.

Since the core now calls BIPStatus(), we don't need to call it
manually in the alert test code anymore (but we do need to make sure
the global bipStatus has the same parameters as our test ones).

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>