Script: Val64: class for efficient access of a stack object.
An unsigned char vector is conveniently aligned for 64-bit access, which is legal in C++ and makes operations much faster.  This implements a wrapper class for safely accessing these, including handling the unaligned case (with a warning to stderr once if this happens, as it never does) and the trailing bytes.

In practice, unaligned accesses are fine on x86 hardware, but the C++ ABI does not support it so we play it safe.  Similarly, access beyond the array end will be hardware safe on all platforms, but we are careful there too.

Note: this is not a generic 64-bit vector class!  It is specifically written with the explicit operations we wish to support.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>