op_div: use Knuth's lookahead trick.
This is faster *and* far more consistent.  To be fair, Brent and Zimmerman
point out that modern branch prediction struggles with the inner underflow
since it happens 50% of the time: the Knuth lookahead helps this a lot.

Doing 100000x10000 byte multiples with Brent & Zimmerman BasecaseDivRem,
we get min-max (avg/stddev):

	Time (ns): 80531099-503553287 (2.41922e+08+/-1e+08)
	Ops per second: 1.990000-12.420000(5.1051+/-2.5)

Using Knuth's lookahead:

	Time (ns): 79143393-79630636 (7.93325e+07+/-9.8e+04)
	Ops per second: 12.560000-12.640000 (12.6049+/-0.016)

Note: we still need to *test* the underflow case, using specific
values.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>